+++ lib/step/steps/rp_pain_slider_question_step.dart
@@ -1,7 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 import 'package:research_package/research_package.dart';
 
-import 'RPUIPainSliderQuestionStep.dart';
+import '../../ui/rpui_pain_slider_question_step.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake, includeIfNull: false)
 class RPPainSliderQuestionStep extends RPStep {
